ml/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
pydantic==1.10.6
bcrypt==4.0.1
google-cloud-firestore==2.8.0
google-cloud-vision==3.0.2
google-auth==2.15.0
spacy==3.4.1
easyocr==1.5.1
pytesseract==0.3.9
pillow==9.2.0
numpy==1.23.4
opencv-python==4.6.0.66
mediapipe==0.8.10
joblib==1.2.0
pandas==1.5.3
regex==2022.10.31
python-dotenv==0.21.0
